moving product files to models folder
