+++ src/main/java/org/soniakbew/TestApplication.java
@@ -5,7 +5,7 @@ import io.dropwizard.setup.Bootstrap;
 import io.dropwizard.setup.Environment;
 import io.federecio.dropwizard.swagger.SwaggerBundle;
 import io.federecio.dropwizard.swagger.SwaggerBundleConfiguration;
-import org.soniakbew.controllers.TestController;
+import org.soniakbew.models.TestController;
 import org.soniakbew.daos.TestDao;
 import org.soniakbew.services.TestService;
